fixes from upstream

+++ install.sh
@@ -17,7 +17,7 @@ info() { printf "\e[0;32m[INFO]\e[m %s\n" "${1}"; exit 0; }
 error() { printf "\e[0;31m[ERROR]\e[m %s\n" "${1}"; exit 1; }
 
 ostree() {
-	if command -v rpm-ostree >/dev/null 2>&1; then
+	if command -v rpm-ostree >/dev/null 2>&1 && [ -f /run/ostree-booted ]; then
 		if curl -fsSL "${_rpm}" -o "/tmp/${_rpm_name}"; then
 			if rpm -qi linuxtoys >/dev/null 2>&1; then
 				if ! sudo rpm-ostree remove linuxtoys; then
@@ -92,6 +92,8 @@ osarch() {
 	if pacman -Qi linuxtoys-bin &>/dev/null; then
 		sudo pacman -R --noconfirm linuxtoys-bin || error "Failed to remove existing linuxtoys-bin package."
 	fi
+	{ pacman -Qi debugedit &>/dev/null || sudo pacman -S debugedit; } || error "Failed to install makepkg dependency debugedit"
+	{ pacman -Qi fakeroot &>/dev/null || sudo pacman -S fakeroot; } || error "Failed to install makepkg dependency fakeroot"
 	mkdir -p ${_pkg_dir}
 	if curl -fsSL "${_pkg}" -o "${_pkg_dir}${_pkg_name}"; then
 		cd "${_pkg_dir}"
